Separated vertex from main
