+++ source/renderer/structs/vertex.hpp
@@ -1,0 +1,85 @@
+/*
+------------------------------------
+
+Slimemaid Source Code (28.04.2022)
+
+This file is part of Slimemaid Source Code.
+
+This file has heading of allocation functions for Vulkan API.
+
+------------------------------------
+*/
+
+#ifndef SM_VULKAN_VERTEX_STRUCT_
+#define SM_VULKAN_VERTEX_STRUCT_
+
+#include <vulkan/vulkan.h>
+
+#define GLM_FORCE_RADIANS
+#define GLM_ENABLE_EXPERIMENTAL
+#define GLM_FORSE_DEPTH_ZERO_TO_ONE
+#define GLM_FORCE_DEFAULT_ALIGNED_GENTYPES // dif
+#include <glm/glm.hpp>
+#include <glm/gtc/matrix_transform.hpp>
+#include <glm/gtx/hash.hpp>
+
+/*
+------------
+Struct describe dot in virtual Vulkan space by position vector and color vector (RGB) and texture coordinate vector.
+Struct has two static methods:
+ - getBindingDescription(): VkVertexInputBindingDescription
+ - getAttributeDescriptions(): array<VkVertexInputAttributeDescription, 3>
+------------
+*/
+struct Vertex {
+  glm::vec3 pos;
+  glm::vec3 color;
+  glm::vec2 texCoord;
+
+  /*
+  ------------
+  Static method return vertex binding description.
+  ------------
+  */
+  static VkVertexInputBindingDescription getBindingDescription() {
+    VkVertexInputBindingDescription bindingDescription{};
+    bindingDescription.binding = 0;
+    bindingDescription.stride = sizeof(Vertex);
+    bindingDescription.inputRate = VK_VERTEX_INPUT_RATE_VERTEX;
+
+    return bindingDescription;
+  }
+
+  /*
+  ------------
+  Static method create and return array for each vertex attribure. Each attribure of array is vertex vector description.
+  ------------
+  */
+  static std::array<VkVertexInputAttributeDescription, 3> getAttributeDescriptions() {
+    std::array<VkVertexInputAttributeDescription, 3> attributeDescriptions{};
+
+    attributeDescriptions[0].binding = 0;
+    attributeDescriptions[0].location = 0;
+    attributeDescriptions[0].format = VK_FORMAT_R32G32B32_SFLOAT;
+    attributeDescriptions[0].offset = offsetof(Vertex, pos);
+
+    attributeDescriptions[1].binding = 0;
+    attributeDescriptions[1].location = 1;
+    attributeDescriptions[1].format = VK_FORMAT_R32G32B32_SFLOAT;
+    attributeDescriptions[1].offset = offsetof(Vertex, color);
+
+    attributeDescriptions[2].binding = 0;
+    attributeDescriptions[2].location = 2;
+    attributeDescriptions[2].format = VK_FORMAT_R32G32_SFLOAT;
+    attributeDescriptions[2].offset = offsetof(Vertex, texCoord);
+
+    return attributeDescriptions;
+  }
+
+  bool operator==(const Vertex &other) const {
+    return pos == other.pos && color == other.color && texCoord == other.texCoord;
+  }
+};
+
+
+#endif
